signals-bot/requirements.txt
discord.py>=2.3.2
tradingview_ta>=3.3.0
yfinance>=0.2.40
pandas>=2.2.2
beautifulsoup4>=4.12.3
lxml>=5.2.2
python-dotenv>=1.0.1
pytz>=2024.1
requests>=2.32.3
matplotlib>=3.7.0
psycopg[binary]>=3.1.18
psycopg-pool>=3.2.0
